+++ Src/search.h
@@ -29,17 +29,17 @@ class Search
         std::vector<std::vector<Node>> close;
         std::list<Node> open;
 
-        int Heuristic(Coordinates, Coordinates, const EnvironmentOptions&) const;
+        double Heuristic(Coordinates, Coordinates, const EnvironmentOptions&) const;
 
-        int Euclidean(Coordinates, Coordinates) const;
+        double Euclidean(Coordinates, Coordinates) const;
 
-        int Manhattan(Coordinates, Coordinates) const;
+        double Manhattan(Coordinates, Coordinates) const;
 
-        int Chebyshev(Coordinates, Coordinates) const;
+        double Chebyshev(Coordinates, Coordinates) const;
 
-        int Diagonal(Coordinates, Coordinates) const;
+        double Diagonal(Coordinates, Coordinates) const;
 
-        int HeuristicWeight() const;
+        double HeuristicWeight() const;
 
         std::optional<Node> GetNeighbours(Node&, int, int, const Map &Map, const EnvironmentOptions &options);
 };
